requirements.txt
python-dotenv==1.0.0; python_version >= "3.8"
pytz==2023.3.post1
